semantic probability

+++ database/migrations/2023_05_25_144710_create_semantic_classes_table.php
@@ -15,7 +15,7 @@ class CreateSemanticClassesTable extends Migration
     {
         Schema::create('semantic_classes', function (Blueprint $table) {
             $table->id();
-            $table->integer('class_id');
+            $table->bigInteger('class_id');
             $table->string('class_name')->nullable()->comment('название класса');
             $table->string('class_probability')->nullable()->comment('вероятность с которой текст принадлежит этому классу');
             $table->timestamps();
